+++ src/idc/reader/imgseg/__init__.py
@@ -1,4 +1,5 @@
 from ._bluechannel import BlueChannelImageSegmentationReader
 from ._grayscale import GrayscaleImageSegmentationReader
 from ._indexedpng import IndexedPngImageSegmentationReader
+from ._instancepng import InstancePngImageSegmentationReader
 from ._layersegments import LayerSegmentsImageSegmentationReader