+++ src/idc/reader/objdet/__init__.py
@@ -1,5 +1,6 @@
 from ._adams import AdamsObjectDetectionReader
 from ._coco import COCOObjectDetectionReader
+from ._instancepng import InstancePngObjectDetectionReader
 from ._opex import OPEXObjectDetectionReader
 from ._roi import ROIObjectDetectionReader
 from ._voc import VOCObjectDetectionReader